+++ lib/feature/presentation/ui/widgets/filter/selected_filters_view.dart.dart
@@ -1,14 +1,16 @@
-import 'package:cportal_flutter/common/custom_theme.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:cportal_flutter/common/util/padding.dart';
 import 'package:cportal_flutter/feature/domain/entities/filter_entity.dart';
 import 'package:cportal_flutter/feature/presentation/bloc/filter_bloc/filter_bloc.dart';
 import 'package:cportal_flutter/feature/presentation/bloc/filter_bloc/filter_state.dart';
-import 'package:cportal_flutter/feature/presentation/ui/widgets/filter/tag_container.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:cportal_flutter/feature/presentation/ui/widgets/filter/filter_view_row.dart';
 
 class SelectedFiltersView extends StatelessWidget {
-  final Function(FilterItemEntity, int) onRemove;
+  final Function(FilterItemEntity, int) onRemove
+  ;
+
+  /// Рендеринг всех выбранных фильтров.
   const SelectedFiltersView({
     Key? key,
     required this.onRemove,
@@ -83,55 +85,5 @@ class SelectedFiltersView extends StatelessWidget {
     });
 
     return isActive;
-  }
-}
-
-class FilterViewRow extends StatelessWidget {
-  final String headline;
-  final List<FilterItemEntity> selectedItems;
-  final Function(FilterItemEntity, int) onClose;
-
-  const FilterViewRow({
-    Key? key,
-    required this.headline,
-    required this.selectedItems,
-    required this.onClose,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final CustomTheme theme = Theme.of(context).extension<CustomTheme>()!;
-
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Padding(
-          padding: const EdgeInsets.only(top: 6),
-          child: Text(
-            headline,
-            style: theme.textTheme.px12.copyWith(
-              color: theme.textLight,
-            ),
-          ),
-        ),
-        const SizedBox(width: 8),
-        Expanded(
-          child: Wrap(
-            spacing: 4,
-            runSpacing: 4,
-            children: List.generate(
-              selectedItems.length,
-              (index) => TagContainer(
-                text: selectedItems[index].name,
-                isCloseAction: true,
-                onTap: () {
-                  onClose(selectedItems[index], index);
-                },
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
   }
 }
